+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 **Updated for Hearthstone 31.4**
 
 **Hearthstone**:
-- Added related card tooltips to Chatty Macaw, Ra-Den, Tram Heist, Joymancer Jepetto, Mixtape, Fate Splitter. \
+- Added related card tooltips to Chatty Macaw, Ra-Den, Tram Heist, Joymancer Jepetto, Mixtape, Fate Splitter, Starlight Whelp. \
   *We will continue adding more tooltips in the coming updates*
 - Added counter for Alien Encounters.
 - Changed when Sasquawk related cards are updated, now it updates when the turn ends so you can have the information to plan your next turn.